Refactor public API and handle some errors properly
