+++ config_struct_test/build.rs
@@ -6,6 +6,9 @@ fn main()
     use config_struct::toml_parsing;
     use config_struct::yaml_parsing;
 
-    toml_parsing::create_module_from_config("config.toml", "src/config/toml.rs");
-    yaml_parsing::create_module_from_config("config.yaml", "src/config/yaml.rs");
+    let toml_config = toml_parsing::parse_config_from_file("config.toml").unwrap();
+    config_struct::write_config_module(&toml_config, "src/config/toml.rs").unwrap();
+
+    let yaml_config = yaml_parsing::parse_config_from_file("config.yaml").unwrap();
+    config_struct::write_config_module(&yaml_config, "src/config/yaml.rs").unwrap();
 }
